test(codex): keep approval collision coverage scoped
